book详情页面的itemcell 高清显示问题修复，历史界面中的星星图标的高清显示的问题解决

+++ Classes/RatingDisplayView.h
@@ -16,9 +16,9 @@
 	
 	NSInteger maxRating;
 	NSInteger minRating;
-	NSInteger rating;
+	CGFloat rating;
 }
-@property(nonatomic,assign) NSInteger rating;
+@property(nonatomic,assign) CGFloat rating;
 
 - (CGRect)rectForStarAtIndex:(NSInteger)index;
 @end
